[ANI-28] refactor: 변경된 폴더구조를 적용합니다.
